fix(catalog-backend): move batch write to EntitiesCatalog where it belongs
