+++ plugins/catalog-backend/src/ingestion/HigherOrderOperations.ts
@@ -14,17 +14,8 @@
  * limitations under the License.
  */
 
-import { ConflictError, InputError } from '@backstage/backend-common';
-import {
-  Entity,
-  entityHasChanges,
-  getEntityName,
-  Location,
-  LocationSpec,
-  serializeEntityRef,
-} from '@backstage/catalog-model';
-import { chunk, groupBy } from 'lodash';
-import limiterFactory from 'p-limit';
+import { InputError } from '@backstage/backend-common';
+import { Entity, Location, LocationSpec } from '@backstage/catalog-model';
 import { v4 as uuidv4 } from 'uuid';
 import { Logger } from 'winston';
 import { EntitiesCatalog, LocationsCatalog } from '../catalog';
@@ -34,25 +25,6 @@ import {
   HigherOrderOperation,
   LocationReader,
 } from './types';
-
-type BatchContext = {
-  kind: string;
-  namespace: string;
-  location: Location;
-};
-
-// Some locations return tens or hundreds of thousands of entities. To make
-// those payloads more manageable, we break work apart in batches of this
-// many entities and write them to storage per batch.
-const BATCH_SIZE = 100;
-
-// When writing large batches, there's an increasing chance of contention in
-// the form of conflicts where we compete with other writes. Each batch gets
-// this many attempts at being written before giving up.
-const BATCH_ATTEMPTS = 3;
-
-// The number of batches that may be ongoing at the same time.
-const BATCH_CONCURRENCY = 3;
 
 /**
  * Placeholder for operations that span several catalogs and/or stretches out
@@ -192,159 +164,35 @@ export class HigherOrderOperations implements HigherOrderOperation {
 
     startTimestamp = process.hrtime();
 
-    await this.batchAddOrUpdateEntities(
-      readerOutput.entities.map(e => e.entity),
-      location,
-    );
+    try {
+      await this.entitiesCatalog.batchAddOrUpdateEntities(
+        readerOutput.entities.map(e => e.entity),
+        location.id,
+      );
+    } catch (e) {
+      for (const entity of readerOutput.entities) {
+        await this.locationsCatalog.logUpdateFailure(
+          location.id,
+          e,
+          entity.entity.metadata.name,
+        );
+      }
+      throw e;
+    }
+
+    this.logger.info(`Posting update success markers`);
+
+    for (const entity of readerOutput.entities) {
+      await this.locationsCatalog.logUpdateSuccess(
+        location.id,
+        entity.entity.metadata.name,
+      );
+    }
 
     this.logger.info(
       `Wrote ${readerOutput.entities.length} entities from location ${
         location.type
       }:${location.target} in ${durationText(startTimestamp)}`,
-    );
-  }
-
-  /**
-   * Writes a number of entities efficiently to storage.
-   *
-   * @param entities Some entities
-   * @param location The location that they all belong to
-   */
-  async batchAddOrUpdateEntities(entities: Entity[], location: Location) {
-    // Group the entities by unique kind+namespace combinations
-    const entitiesByKindAndNamespace = groupBy(entities, entity => {
-      const name = getEntityName(entity);
-      return `${name.kind}:${name.namespace}`.toLowerCase();
-    });
-
-    const limiter = limiterFactory(BATCH_CONCURRENCY);
-    const tasks: Promise<void>[] = [];
-
-    for (const groupEntities of Object.values(entitiesByKindAndNamespace)) {
-      const { kind, namespace } = getEntityName(groupEntities[0]);
-
-      // Go through the new entities in reasonable chunk sizes (sometimes,
-      // sources produce tens of thousands of entities, and those are too large
-      // batch sizes to reasonably send to the database)
-      for (const batch of chunk(groupEntities, BATCH_SIZE)) {
-        tasks.push(
-          limiter(async () => {
-            const first = serializeEntityRef(batch[0]);
-            const last = serializeEntityRef(batch[batch.length - 1]);
-            this.logger.debug(
-              `Considering batch ${first}-${last} (${batch.length} entries)`,
-            );
-
-            // Retry the batch write a few times to deal with contention
-            const context = { kind, namespace, location };
-            for (let attempt = 1; attempt <= BATCH_ATTEMPTS; ++attempt) {
-              try {
-                const { toAdd, toUpdate } = await this.analyzeBatch(
-                  batch,
-                  context,
-                );
-                if (toAdd.length) await this.batchAdd(toAdd, context);
-                if (toUpdate.length) await this.batchUpdate(toUpdate, context);
-                break;
-              } catch (e) {
-                if (e instanceof ConflictError && attempt < BATCH_ATTEMPTS) {
-                  this.logger.warn(
-                    `Failed to write batch at attempt ${attempt}/${BATCH_ATTEMPTS}, ${e}`,
-                  );
-                } else {
-                  throw e;
-                }
-              }
-            }
-          }),
-        );
-      }
-    }
-
-    await Promise.all(tasks);
-  }
-
-  // Given a batch of entities that were just read from a location, take them
-  // into consideration by comparing against the existing catalog entities and
-  // produce the list of entities to be added, and the list of entities to be
-  // updated
-  private async analyzeBatch(
-    newEntities: Entity[],
-    { kind, namespace }: BatchContext,
-  ): Promise<{
-    toAdd: Entity[];
-    toUpdate: Entity[];
-  }> {
-    const markTimestamp = process.hrtime();
-
-    const oldEntities = await this.entitiesCatalog.entities({
-      kind: kind,
-      'metadata.namespace': namespace,
-      'metadata.name': newEntities.map(e => e.metadata.name),
-    });
-
-    const oldEntitiesByName = new Map(
-      oldEntities.map(e => [e.metadata.name, e]),
-    );
-
-    const toAdd: Entity[] = [];
-    const toUpdate: Entity[] = [];
-
-    for (const newEntity of newEntities) {
-      const oldEntity = oldEntitiesByName.get(newEntity.metadata.name);
-      if (!oldEntity) {
-        toAdd.push(newEntity);
-      } else if (entityHasChanges(oldEntity, newEntity)) {
-        toUpdate.push(newEntity);
-      }
-    }
-
-    this.logger.debug(
-      `Found ${toAdd.length} entities to add, ${
-        toUpdate.length
-      } entities to update in ${durationText(markTimestamp)}`,
-    );
-
-    return { toAdd, toUpdate };
-  }
-
-  // Efficiently adds the given entities to storage, under the assumption that
-  // they do not conflict with any existing entities
-  private async batchAdd(entities: Entity[], { location }: BatchContext) {
-    const markTimestamp = process.hrtime();
-
-    await this.entitiesCatalog.addEntities(entities, location.id);
-
-    // TODO(freben): Still not batched
-    for (const entity of entities) {
-      await this.locationsCatalog.logUpdateSuccess(
-        location.id,
-        entity.metadata.name,
-      );
-    }
-
-    this.logger.debug(
-      `Added ${entities.length} entities in ${durationText(markTimestamp)}`,
-    );
-  }
-
-  // Efficiently updates the given entities into storage, under the assumption
-  // that there already exist entities with the same names
-  private async batchUpdate(entities: Entity[], { location }: BatchContext) {
-    const markTimestamp = process.hrtime();
-
-    // TODO(freben): Still not batched
-    for (const entity of entities) {
-      await this.entitiesCatalog.addOrUpdateEntity(entity);
-
-      await this.locationsCatalog.logUpdateSuccess(
-        location.id,
-        entity.metadata.name,
-      );
-    }
-
-    this.logger.debug(
-      `Updated ${entities.length} entities in ${durationText(markTimestamp)}`,
     );
   }
 }
